+++ example/lib/main.dart
@@ -33,18 +33,46 @@ class _MyHomePageState extends State<MyHomePage> {
 }
 
 class Messaging {
-  static String token;
+  static String? token;
+
+  static deleteToken() {
+    Messaging.token = null;
+    FCM.deleteRefreshToken();
+  }
+
+  static subscribeToTopic(String topic) {
+    FCM.subscribeToTopic(topic);
+  }
+
+  static unsubscribeFromTopic(String topic) {
+    FCM.unsubscribeFromTopic(topic);
+  }
+
+  @pragma('vm:entry-point')
+  static Future<void> onNotificationReceived(RemoteMessage message) async {
+    await Firebase.initializeApp();
+    //print('Handling a message ${message}');
+  }
+
+  @pragma('vm:entry-point')
   static initFCM() async {
     try {
+      await Firebase.initializeApp();
       await FCM.initializeFCM(
-          onNotificationPressed: (Map<String, dynamic> data) {
-            print(data);
-          },
-          onTokenChanged: (String token) {
+        withLocalNotification: true,
+        // navigatorKey: Keys.navigatorKey,
+        onNotificationReceived: onNotificationReceived,
+        onNotificationPressed: (Map<String, dynamic> data) {},
+        onTokenChanged: (String? token) {
+          if (token != null) {
             Messaging.token = token;
-            print(token);
-          },
-          icon: 'icon');
-    } catch (e) {}
+          }
+        },
+        // TODO add this icon to android/app/src/main/res/drawable/ic_launcher.png
+        icon: 'ic_launcher',
+      );
+    } catch (e) {
+      print(e);
+    }
   }
 }