+++ lib/notification/fcm.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
@@ -8,9 +7,20 @@ import 'local_notification.dart';
 class FCM {
   FCM._();
 
+  /// Callback for token changes.
   static late ValueChanged<String?> _onTokenChanged;
 
-  static initializeFCM(
+  /// Initialize Firebase Cloud Messaging and set up notification handling.
+  ///
+  /// [onTokenChanged] is a required callback function that is invoked when the FCM token changes.
+  /// [onNotificationPressed] is an optional callback to handle notifications when pressed by the user.
+  /// [onNotificationReceived] is a background message handler called when a notification is received while the app is in the background.
+  /// [navigatorKey] is an optional `GlobalKey<NavigatorState>` for navigating within the app.
+  /// [icon] is a string specifying the icon used for displaying notifications. icon must be in android/app/src/main/res/drawable/ic_launcher.png
+  /// [withLocalNotification] is a boolean flag to enable or disable local notifications.
+  ///
+  /// This method initializes Firebase, sets up token handling, background message handling, and notification presentation options for iOS and Android.
+  static Future<void> initializeFCM(
       {required void onTokenChanged(String? token),
       void onNotificationPressed(Map<String, dynamic> data)?,
       required BackgroundMessageHandler onNotificationReceived,
@@ -18,32 +28,42 @@ class FCM {
       required String icon,
       bool withLocalNotification = true}) async {
     _onTokenChanged = onTokenChanged;
+    await Firebase.initializeApp();
+
+    FirebaseMessaging messaging = FirebaseMessaging.instance;
+
+    NotificationSettings settings = await messaging.requestPermission(
+      alert: true,
+      announcement: false,
+      badge: true,
+      carPlay: false,
+      criticalAlert: false,
+      provisional: false,
+      sound: true,
+    );
+
+    print('User granted permission: ${settings.authorizationStatus}');
+
     await LocalNotification.initializeLocalNotification(
         onNotificationPressed: onNotificationPressed, icon: icon);
-    await Firebase.initializeApp();
-    FirebaseMessaging.instance.getToken().then(onTokenChanged);
-    Stream<String> _tokenStream = FirebaseMessaging.instance.onTokenRefresh;
+    messaging.getToken().then(onTokenChanged);
+    Stream<String> _tokenStream = messaging.onTokenRefresh;
     _tokenStream.listen(onTokenChanged);
 
     // Set the background messaging handler early on, as a named top-level function
     FirebaseMessaging.onBackgroundMessage(onNotificationReceived);
 
-    /// Update the iOS foreground notification presentation options to allow
-    /// heads up notifications.
-    await FirebaseMessaging.instance
-        .setForegroundNotificationPresentationOptions(
+    await messaging.setForegroundNotificationPresentationOptions(
       alert: true,
       badge: true,
       sound: true,
     );
 
-    FirebaseMessaging.instance
-        .getInitialMessage()
-        .then((RemoteMessage? message) {
+    messaging.getInitialMessage().then((RemoteMessage? message) {
       print('getInitialMessage');
       print(message);
       if (message != null) {
-        if (navigatorKey != null)
+        if (navigatorKey != null) {
           Timer.periodic(
             Duration(milliseconds: 500),
             (timer) {
@@ -52,6 +72,7 @@ class FCM {
               timer.cancel();
             },
           );
+        }
       }
     });
 
@@ -80,9 +101,19 @@ class FCM {
     });
   }
 
-  //static Future<void> _firebaseMessagingBackgroundHandler
+  /// Delete the FCM refresh token and retrieve a new token.
   static deleteRefreshToken() {
     FirebaseMessaging.instance.deleteToken();
     FirebaseMessaging.instance.getToken().then(_onTokenChanged);
+  }
+
+  /// To subscribe to a topic, call subscribeToTopic() with the topic name. This method returns a Future, which resolves when the subscription succeeded:
+  static subscribeToTopic(String topic) {
+    FirebaseMessaging.instance.subscribeToTopic(topic);
+  }
+
+  /// To unsubscribe from a topic, call the unsubscribeFromTopic method with the topic name:
+  static unsubscribeFromTopic(String topic) {
+    FirebaseMessaging.instance.unsubscribeFromTopic(topic);
   }
 }